+++ lib/exception_handler/exception_handler.h
@@ -1,0 +1,14 @@
+#ifndef EXCEPTION_HANDLER_H
+#define EXCEPTION_HANDLER_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void setupGlobalExceptionHandler();
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // EXCEPTION_HANDLER_H
