+++ platformio.ini
@@ -12,7 +12,7 @@
 [platformio]
 name = "Chelonian Access"
 description = "Modern RFID Entry System - Shelled in Security"
-default_envs = esp32c3, native
+default_envs = esp32c3
 
 ; Common settings for all environments
 [env]
@@ -64,17 +64,12 @@ build_flags =
     -D USINGMP3
     -I lib/mocks
     -I lib/test_helpers
-    -I lib/access_service
-    -I lib/audio_controller
-    -I lib/relay_controller
-    -I lib/rfid_controller
     -lgcov
     --coverage             ; Enable coverage
     -ftest-coverage        ; Add test coverage
 build_unflags = -Os        ; Disable optimization for accurate coverage
 lib_deps =
     Unity
-build_src_filter = +<test/> +<lib/mocks/> -<*> ; Include test and mock source files
 test_ignore =
     test_embedded
     test_native
